+++ src/commonMain/kotlin/com/hoc081098/flowext/utils/NULL_VALUE.kt
@@ -35,13 +35,3 @@ import kotlin.jvm.JvmField
 @DelicateFlowExtApi
 @JvmField
 public val NULL_VALUE: Symbol = Symbol("NULL_VALUE")
-
-/**
- * This is a work-around for having nested nulls in generic code.
- * This allows for writing faster generic code instead of using `Option`.
- * This is only used as an optimisation technique in low-level code.
- *
- * This is internal and should not be used outside of the library.
- */
-@JvmField
-internal val INTERNAL_NULL_VALUE: Symbol = Symbol("INTERNAL_NULL_VALUE")