+++ src/commonTest/kotlin/com/hoc081098/flowext/FlowFromNonSuspendTest.kt
@@ -32,20 +32,16 @@ import com.hoc081098.flowext.utils.test
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.take
-import kotlinx.coroutines.withContext
 
 @ExperimentalCoroutinesApi
-class FlowFromSuspendTest : BaseTest() {
+class FlowFromNonSuspendTest : BaseTest() {
   @Test
-  fun flowFromSuspendEmitsValues() = runTest {
+  fun flowFromNonSuspendEmitsValues() = runTest {
     var count = 0L
-    val flow = flowFromSuspend {
-      delay(count)
+    val flow = flowFromNonSuspend {
       count
     }
 
@@ -59,17 +55,17 @@ class FlowFromSuspendTest : BaseTest() {
   }
 
   @Test
-  fun flowFromSuspendFunctionThrows() = runTest {
+  fun flowFromNonSuspendFunctionThrows() = runTest {
     val testException = TestException()
 
-    flowFromSuspend<Int> { throw testException }.test(listOf(Event.Error(testException)))
+    flowFromNonSuspend<Int> { throw testException }.test(listOf(Event.Error(testException)))
   }
 
   @Test
-  fun flowFromSuspendCancellation() = runTest {
+  fun flowFromNonSuspendCancellation() = runTest {
     fun throwsCancellationException(): Unit = throw CancellationException("Flow was cancelled")
     assertFailsWith<CancellationException>(
-      flowFromSuspend {
+      flowFromNonSuspend {
         val i = 1 + 2
         throwsCancellationException()
         i + 3
@@ -78,43 +74,17 @@ class FlowFromSuspendTest : BaseTest() {
   }
 
   @Test
-  fun flowFromSuspendTake() = runTest {
-    flowFromSuspend { 100 }.take(1)
-      .test(listOf(Event.Value(100), Event.Complete))
-
-    flowFromSuspend {
-      delay(100)
-      100
-    }
+  fun flowFromNonSuspenddTake() = runTest {
+    flowFromNonSuspend { 100 }
       .take(1)
       .test(listOf(Event.Value(100), Event.Complete))
   }
 
   @Test
   fun testContextPreservation1() = runTest {
-    val flow = flowFromSuspend {
-      assertEquals("OK", NamedDispatchers.name())
-
-      withContext(Dispatchers.Default) { delay(100) }
-
+    val flow = flowFromNonSuspend {
       assertEquals("OK", NamedDispatchers.name())
       42
-    }.flowOn(NamedDispatchers("OK"))
-
-    flow.test(listOf(Event.Value(42), Event.Complete))
-  }
-
-  @Test
-  fun testContextPreservation2() = runTest {
-    val flow = flowFromSuspend {
-      assertEquals("OK", NamedDispatchers.name())
-
-      withContext(Dispatchers.Default) {
-        delay(100)
-        42
-      }.also {
-        assertEquals("OK", NamedDispatchers.name())
-      }
     }.flowOn(NamedDispatchers("OK"))
 
     flow.test(listOf(Event.Value(42), Event.Complete))